pyproject.toml
[project]
name = "seqrepo-rest-service"
authors = [
    { name = "biocommons contributors", email = "biocommons-dev@googlegroups.com" },
]
description = "SeqRepo REST Service"
readme = "README.md"
license = { file="LICENSE.txt" }
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dynamic = ["version"]
dependencies = [
    "biocommons.seqrepo ~= 0.6",
    "coloredlogs",
    "connexion[swagger-ui] ~= 2.2",
    "Flask ~= 2.2",
]

[project.optional-dependencies]
dev = [
    "bandit ~= 1.7",
    "build ~= 0.8",
    "flake8 ~= 4.0",
    "ipython ~= 8.4",
    "isort ~= 5.10",
    "mypy-extensions ~= 1.0",
    "pre-commit ~= 3.4",
    "pylint ~= 2.14",
    "pytest-cov ~= 4.1",
    "pytest-optional-tests",
    "pytest ~= 7.1",
    "pyright~=1.1",
    "ruff == 0.4.4",
    "tox ~= 3.25",
    "vcrpy",
]
docs = ["mkdocs"]

[project.scripts]
seqrepo-rest-service = "seqrepo_rest_service.cli:main"

[project.urls]
"Homepage" = "https://github.com/biocommons/seqrepo-rest-service"
"Bug Tracker" = "https://github.com/biocommons/seqrepo-rest-service/issues"

[build-system]
requires = ["setuptools ~= 69.0", "setuptools_scm[toml] ~= 8.0"]
build-backend = "setuptools.build_meta"


[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["__pycache__", "*.pyc"]
namespaces = true

[tool.setuptools.package-data]
"*" = ["*.gz", "*.json", "*.yaml"]


[tool.setuptools_scm]


# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
[tool.pytest.ini_options]
addopts = "-s -v -x --strict-markers -m 'not extra' --doctest-modules --cov=src"
doctest_optionflags = [
    "ALLOW_BYTES",
    "ALLOW_UNICODE",
    "ELLIPSIS",
    "IGNORE_EXCEPTION_DETAIL",
    "NORMALIZE_WHITESPACE",
]
markers = [
    "network: tests that require network connectivity",
    "slow: slow tests that should be run infrequently",
    "vcr: tests with cached data",
]


[tool.coverage.run]
branch = true
source = ["biocommons.example"]
omit = ["*_test.py", "*/test/*", "*/tests/*"]


[tool.coverage.report]
show_missing = true
exclude_lines = [
    # Have to re-enable the standard pragma
    "pragma: no cover",

    # Don't complain about missing debug-only code:
    "def __repr__",
    "if self.debug",

    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",

    # Don't complain if non-runnable code isn't run:
    "if __name__ == .__main__.:",
]

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.pyright]
include = ["src", "tests"]


[tool.pylint.'MESSAGES CONTROL']
disable = "R0913"


[tool.pylint.format]
max-line-length = 100

[tool.ruff]
src = ["src", "tests"]
line-length = 100

